+++ src/callbacks.h
@@ -26,3 +26,7 @@ void nvs_clear_cancelled_callback(void* context);
 void show_app_info(void* context);
 void app_info_ok_callback(void* context);
 void app_info_cancel_callback(void* context);
+void wardrive_clear_confirmed_callback(void* context);
+void wardrive_clear_cancelled_callback(void* context);
+void pcap_clear_confirmed_callback(void* context);
+void pcap_clear_cancelled_callback(void* context);